chore: v2 refactor part2
