--- RClick/Shared/ AppStore.swift
@@ -1,15 +1,0 @@
-//
-//   AppStore.swift
-//  RClick
-//
-//  Created by 李旭 on 2024/8/9.
-//
-
-import Foundation
-import Observation
-
-@Observable
-final class AppStore: Identifiable {
-    var extensionEnable = false
-}
-
